+++ src/grab.rs
@@ -312,12 +312,13 @@ fn score_elements<'a>(
     flags: &FlagSet<GrabFlags>,
 ) -> Vec<NodeRef<'a>> {
     let mut score_map: HashMap<NodeId, f32> = HashMap::default();
+    let mut cc_cache = CharCounterCache::default();
 
     for element in elements_to_score {
         if element.parent().is_none() {
             continue;
         }
-        let content_len = element.normalized_char_count();
+        let content_len = cc_cache.char_count(element);
         if content_len < 25 {
             continue;
         }
@@ -360,6 +361,8 @@ fn score_elements<'a>(
         }
     }
 
+    
+
     // Scale the final candidates score based on link density. Good content
     // should have a relatively small link density (5% or less) and be mostly
     // unaffected by this operation.
@@ -371,7 +374,7 @@ fn score_elements<'a>(
             let candidate = NodeRef::new(node_id, tree);
             // Skipping adjustment of low score
             let score = if prev_score > cfg.min_score_to_adjust {
-                prev_score * (1.0 - link_density(&candidate, None))
+                prev_score * (1.0 - link_density_fn(&candidate, None, |n| cc_cache.char_count(n)))
             } else {
                 prev_score
             };